feat(api): move #authenticateApplication controller method to identity access management context
