+++ api/tests/identity-access-management/unit/application/token.controller.test.js
@@ -1,3 +1,4 @@
+import { usecases as libUsecases } from '../../../../lib/domain/usecases/index.js';
 import { tokenController } from '../../../../src/identity-access-management/application/token/token.controller.js';
 import { usecases } from '../../../../src/identity-access-management/domain/usecases/index.js';
 import { RequestedApplication } from '../../../../src/identity-access-management/infrastructure/utils/network.js';
@@ -187,6 +188,54 @@ describe('Unit | Identity Access Management | Application | Controller | Token',
 
       // then
       expect(response).to.be.null;
+    });
+  });
+
+  describe('#authenticateApplication', function () {
+    it('returns an OAuth 2 token response', async function () {
+      // given
+      const access_token = 'jwt.access.token';
+      const client_id = Symbol('clientId');
+      const client_secret = Symbol('clientSecret');
+      const scope = Symbol('scope');
+
+      const request = {
+        headers: {
+          'content-type': 'application/x-www-form-urlencoded',
+        },
+        payload: {
+          grant_type: 'password',
+          client_id,
+          client_secret,
+          scope,
+        },
+      };
+      const tokenServiceStub = { extractUserId: sinon.stub() };
+      tokenServiceStub.extractUserId.returns(client_id);
+
+      const dependencies = { tokenService: tokenServiceStub };
+
+      sinon
+        .stub(libUsecases, 'authenticateApplication')
+        .withArgs({ clientId: client_id, clientSecret: client_secret, scope })
+        .resolves(access_token);
+
+      // when
+      const response = await tokenController.authenticateApplication(request, hFake, dependencies);
+
+      // then
+      const expectedResponseSource = {
+        token_type: 'bearer',
+        access_token,
+        client_id,
+      };
+      expect(response.source).to.deep.equal(expectedResponseSource);
+      expect(response.statusCode).to.equal(200);
+      expect(response.headers).to.deep.equal({
+        'Content-Type': 'application/json;charset=UTF-8',
+        'Cache-Control': 'no-store',
+        Pragma: 'no-cache',
+      });
     });
   });
 });
